pyproject.toml
[project]
name = "gradio-adk-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=1.20.0",
    "gradio>=6.1.0",
]
